Fix exception on home page
